pyproject.toml
[project]
name = "google-adk-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.17.0",
    "litellm>=1.78.7",
    "loguru>=0.7.3",
    "rich>=14.2.0",
    "typer>=0.20.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["google_adk_test*"]
